🔥 remove a couple unnecessary bits
