+++ modules/store/src/wrappers/index.ts
@@ -1,4 +1,3 @@
 export * from "./asyncStorage";
 export * from "./localStorage";
 export * from "./sequelizeStorage";
-export * from "./memoryStorage";
