+++ modules/types/package.json
@@ -15,14 +15,14 @@
     "build": "rm -rf ./dist/* && ./node_modules/.bin/tsc -b . && ./node_modules/.bin/rollup -c",
     "lint": "../../node_modules/.bin/eslint -c '../../.eslintrc.js' --fix 'src/**/*'"
   },
-  "dependencies": {},
+  "dependencies": {
+    "ethers": "4.0.47"
+  },
   "devDependencies": {
-    "ethers": "4.0.47",
     "eventemitter3": "4.0.4",
     "evt": "1.7.11",
     "rollup": "2.12.0",
     "rollup-plugin-typescript2": "0.27.1",
-    "sequelize": "5.21.11",
     "typescript": "3.9.3"
   }
 }
