+++ modules/types/src/store.ts
@@ -1,5 +1,3 @@
-import { Sequelize } from "sequelize";
-
 import { AppInstanceJson, AppInstanceProposal } from "./app";
 import { Address, Bytes32 } from "./basic";
 import {
@@ -7,61 +5,16 @@ import {
   MinimalTransaction,
   SetStateCommitmentJSON,
 } from "./commitments";
-import { ILoggerService } from "./logger";
 import { StateChannelJSON } from "./state";
-import { enumify } from "./utils";
 import { IWatcherStoreService } from "./watcher";
 
 export const ConnextNodeStorePrefix = "INDRA_NODE_CF_CORE";
 export const ConnextClientStorePrefix = "INDRA_CLIENT_CF_CORE";
 
-// TODO: remove StoreTypes during next breaking release
-export const StoreTypes = enumify({
-  AsyncStorage: "AsyncStorage",
-  File: "File",
-  LocalStorage: "LocalStorage",
-  Postgres: "Postgres",
-  Memory: "Memory",
-});
-export type StoreTypes = typeof StoreTypes[keyof typeof StoreTypes];
-
 export type StorePair = {
   path: string;
   value: any;
 };
-
-export interface AsyncStorageData {
-  [key: string]: any;
-}
-
-export type InitCallback = (data: AsyncStorageData) => void;
-
-export interface IAsyncStorage {
-  getItem(key: string): Promise<string | null>;
-  setItem(key: string, value: string): Promise<void>;
-  removeItem(key: string): Promise<void>;
-}
-
-// TODO: Remove
-export interface FileStorageOptions {
-  fileExt?: string;
-  fileDir?: string;
-}
-
-export interface SequelizeStorageOptions {
-  sequelize?: Sequelize | string;
-  dbTableName?: string;
-}
-
-// TODO: Remove
-export interface StoreFactoryOptions extends FileStorageOptions, SequelizeStorageOptions {
-  logger?: ILoggerService;
-  storage?: IAsyncStorage | any;
-  prefix?: string;
-  separator?: string;
-  asyncStorageKey?: string;
-  backupService?: IBackupService;
-}
 
 export interface IBackupService {
   restore(): Promise<StorePair[]>;
@@ -73,10 +26,6 @@ export type WithdrawalMonitorObject = {
   retry: number;
   tx: MinimalTransaction;
 };
-
-export interface ChannelsMap {
-  [multisigAddress: string]: any;
-}
 
 export const STORE_SCHEMA_VERSION = 1;
